requirements.txt
fastapi
uvicorn
google-genai
pydantic
python-dotenv
pytest
python-multipart
pypdf
python-gitlab
qdrant-client
tiktoken
